add pointer to mirror if DL fails
set up a mirror of the superFreq resources on figshare, including all genomes and modes. In case the WEHI gitlab server goes offline. Failed DL will generate an error message pointing to the mirror. User will ahve to manually download in this case, but at lest there is an option. Further instructions at the figshare site.